refactor: demo timer
