+++ include/kernel/cmd.hpp
@@ -11,7 +11,6 @@ int cmd_cat(int, char*[]);
 int cmd_alloc(int, char*[]);
 int cmd_devtree(int, char*[]);
 int cmd_run(int, char*[]);
-int cmd_timer(int, char*[]);
 int cmd_setTimeout(int, char*[]);
 int cmd_uart(int, char*[]);
 int cmd_demo(int, char*[]);
