+++ src/kernel/cmds/demo.cpp
@@ -4,6 +4,36 @@
 #include "string.hpp"
 #include "timer.hpp"
 
+bool show_timer = false;
+int timer_delay = 0;
+
+void print_2s_timer(void* context) {
+  if (timer_delay > 0) {
+    auto delay = timer_delay;
+    timer_delay = 0;
+    mini_uart_printf_sync("delay timer %ds\n", delay);
+    auto cur = get_current_tick();
+    while (get_current_tick() - cur < delay * freq_of_timer)
+      NOP;
+  }
+  if (show_timer) {
+    mini_uart_printf("[" PRTval "] 2s timer interrupt\n",
+                     FTval(get_current_time()));
+    add_timer({2, 0}, (void*)context, (Timer::fp)context);
+  }
+}
+
+int demo_timer(int argc, char* argv[]) {
+  if (show_timer) {
+    show_timer = false;
+  } else {
+    show_timer = true;
+    add_timer({2, 0}, (void*)print_2s_timer, print_2s_timer);
+  }
+  mini_uart_printf("%s timer interrupt\n", show_timer ? "show" : "hide");
+  return 0;
+}
+
 int demo_preempt(int argc, char* argv[]) {
   auto print_data = [](void* ctx) {
     mini_uart_printf("print from task %ld\n", (long)ctx);
@@ -39,6 +69,10 @@ int cmd_demo(int argc, char* argv[]) {
   }
 
   auto cmd = argv[1];
+
+  if (!strcmp(cmd, "timer"))
+    return demo_timer(argc, argv);
+
   if (!strcmp(cmd, "preempt"))
     return demo_preempt(argc, argv);
